requirements.txt
flask==2.3.3
flask-login
Flask-SQLAlchemy
Flask-Migrate
flask-socketio
flask_mail
gunicorn==20.0.4
pillow
psycopg2-binary
python-dotenv
werkzeug==2.3.7
reverse-geocode